+++ extended/src/main/java/apoc/custom/CustomProcedureInfo.java
@@ -11,11 +11,13 @@ import java.util.Optional;
 
 import static apoc.ExtendedSystemLabels.Function;
 import static apoc.ExtendedSystemLabels.Procedure;
-import static apoc.custom.CypherHandlerNewProcedure.getFunctionInfo;
-import static apoc.custom.CypherHandlerNewProcedure.getProcedureInfo;
+import static apoc.custom.CypherProceduresUtil.getFunctionInfo;
+import static apoc.custom.CypherProceduresUtil.getProcedureInfo;
 import static apoc.custom.CypherProceduresHandler.FUNCTION;
 import static apoc.custom.CypherProceduresHandler.PREFIX;
 import static apoc.custom.CypherProceduresHandler.PROCEDURE;
+import static apoc.custom.CypherProceduresUtil.getFunctionInfo;
+import static apoc.custom.CypherProceduresUtil.getProcedureInfo;
 
 public class CustomProcedureInfo {
     public String type;
