+++ extended/src/main/java/apoc/systemdb/metadata/ExportFunction.java
@@ -4,6 +4,7 @@ import apoc.ExtendedSystemPropertyKeys;
 import apoc.SystemPropertyKeys;
 import apoc.custom.CypherHandlerNewProcedure;
 import apoc.custom.CypherProceduresHandler;
+import apoc.custom.CypherProceduresUtil;
 import apoc.export.util.ProgressReporter;
 import org.apache.commons.lang3.tuple.Pair;
 import org.neo4j.graphdb.Node;
@@ -35,7 +36,7 @@ public class ExportFunction implements ExportMetadata {
 
 
     static String getSignature(Node node, String name) {
-        return CypherHandlerNewProcedure.deserializeSignatures((String) node.getProperty(name))
+        return CypherProceduresUtil.deserializeSignatures((String) node.getProperty(name))
                 .stream().map(FieldSignature::toString)
                 .collect(Collectors.joining(", "));
     }
