+++ extended/src/main/java/apoc/systemdb/metadata/ExportProcedure.java
@@ -3,6 +3,7 @@ package apoc.systemdb.metadata;
 import apoc.ExtendedSystemPropertyKeys;
 import apoc.SystemPropertyKeys;
 import apoc.custom.CypherHandlerNewProcedure;
+import apoc.custom.CypherProceduresUtil;
 import apoc.export.util.ProgressReporter;
 import org.apache.commons.lang3.tuple.Pair;
 import org.neo4j.graphdb.Node;
@@ -33,7 +34,7 @@ public class ExportProcedure implements ExportMetadata {
 
 
     static String getSignature(Node node, String name) {
-        return CypherHandlerNewProcedure.deserializeSignatures((String) node.getProperty(name))
+        return CypherProceduresUtil.deserializeSignatures((String) node.getProperty(name))
                 .stream().map(FieldSignature::toString)
                 .collect(Collectors.joining(", "));
     }
